Add Duplex to the printing options and configure the OO printing interface as well as the Jasper interface.
Add Duplex to the printing options and configure the OO printing interface as well as the Jasper interface.
move duplex to the Document Printer relationship

update adl to correct error

Add Duplex to the printing options and configure the OO printing interface as well as the Jasper interface.
move duplex to the Document Printer relationship

Correct Licensing and Finalize changes